+++ main/tools/parser.h
@@ -7,7 +7,7 @@
 
 #include "dwin_common.h"
 
-uint16_t get_color_temp(int8_t temp);
+uint16_t get_color_temp(int temp);
 const char *get_rssi_description(uint8_t rssi);
 const char *get_auth_mode(int authmode);
 picture_t get_pic(const uint weather_id, const bool night);
